// components/types.ts
export interface Evaluation {
    nombres_apellidos: string
    cedula: string
    cargo: string
    fecha_evaluacion: string
    anio: number
    porcentaje_calificacion: number
    acuerdos_mejora_desempeno_colaborador: string
    acuerdos_mejora_desempeno_jefe: string
    necesidades_desarrollo: string
    aspectos_positivos: string
    compromiso_pasion_entrega: number
    compromiso?: number
    honestidad: number
    respeto: number
    sencillez: number
    servicio: number
    trabajo_equipo: number
    conocimiento_trabajo: number
    productividad: number
    cumple_sistema_gestion: number
    total_puntos: number
    departamento: string
    area_jefe_pertenencia: string
  }
  
  